backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
aiohttp==3.9.1
pillow==10.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6
